clean code one line if rename package
