+++ src/main/java/search_engine/query_decoder/QueryDecoder.java
@@ -1,4 +1,4 @@
-package search_engine.queryDecoder;
+package search_engine.query_decoder;
 
 public interface QueryDecoder {
     Query decode(String query);
